Factorisation JourneyController : migration de méthodes métiers (manipulation données géo) vers GeoService
