+++ app/Services/JourneyService.php
@@ -294,6 +294,68 @@ class JourneyService
     }
 
     /**
+     * Filtre une liste de trajets candidats en ne gardant que ceux dont le tracé
+     * passe à proximité du départ ET de l'arrivée recherchés.
+     *
+     * Pour chaque trajet : on récupère les points du tracé GeoJSON, on cherche le
+     * point le plus proche du départ ; s'il est dans le rayon, on cherche le point
+     * le plus proche de l'arrivée PARMI LES POINTS SUIVANTS (pour garantir que le
+     * trajet va bien dans le sens départ → arrivée).
+     *
+     * Filtres partiels gérés : seul le départ, seule l'arrivée, ou aucun des deux
+     * (auquel cas tous les trajets sont retournés).
+     *
+     * @param  array[]    $journeys      Trajets candidats (doivent contenir 'track_id')
+     * @param  array|null $start         Point de départ ['lat' => float, 'lon' => float] ou null
+     * @param  array|null $end           Point d'arrivée ['lat' => float, 'lon' => float] ou null
+     * @param  float      $maxDistanceKm Rayon de tolérance en km (défaut : 10)
+     * @return array[]   Sous-ensemble des trajets correspondants
+     */
+    public function findMatchingJourneys(array $journeys, ?array $start, ?array $end, float $maxDistanceKm = 10): array
+    {
+        if ($start === null && $end === null) {
+            return $journeys;
+        }
+
+        $matchingJourneys = [];
+
+        foreach ($journeys as $journey) {
+
+            $points = $this->geoService->getTrackPoints((int) $journey['track_id']);
+
+            if (empty($points)) {
+                continue;
+            }
+
+            $startIndex = 0;
+
+            // --- Contrainte sur le départ
+            if ($start !== null) {
+                $startIndex      = $this->geoService->findClosestPointIndex($points, $start);
+                $distanceToStart = $this->geoService->calculateDistance($start, $points[$startIndex]);
+
+                if ($distanceToStart > $maxDistanceKm) {
+                    continue;
+                }
+            }
+
+            // --- Contrainte sur l'arrivée (après le départ)
+            if ($end !== null) {
+                $endIndex      = $this->geoService->findClosestPointIndex($points, $end, $startIndex);
+                $distanceToEnd = $this->geoService->calculateDistance($end, $points[$endIndex]);
+
+                if ($distanceToEnd > $maxDistanceKm) {
+                    continue;
+                }
+            }
+
+            $matchingJourneys[] = $journey;
+        }
+
+        return $matchingJourneys;
+    }
+
+    /**
      * Recherche les demandes de trajet compatibles avec un trajet nouvellement créé
      * et envoie un mail de notification à chaque demandeur concerné.
      *
@@ -381,4 +443,32 @@ class JourneyService
             );
         }
     }
+
+    /**
+     * Retourne le trajet existant de l'utilisateur sur la même demi-journée
+     * (matin : 00h–12h, après-midi : 12h–24h) que la date/heure fournies.
+     *
+     * Sert à prévenir la création de doublons par un même conducteur sur
+     * une plage horaire incompatible.
+     *
+     * @param  int    $userId    Identifiant du conducteur
+     * @param  string $startDate Date de départ au format Y-m-d
+     * @param  string $startTime Heure de départ au format H:i
+     * @return array|null Trajet existant sur la demi-journée, ou null si aucun
+     */
+    private function findExistingJourneyOnHalfDay(int $userId, string $startDate, string $startTime): ?array
+    {
+        $journeyStartDate = new DateTimeImmutable($startDate);
+        $journeyStartTime = new DateTimeImmutable($startTime);
+
+        $startHour        = (int) $journeyStartTime->format('H') < 12 ? 0 : 12;
+        $dayStartDateTime = $journeyStartDate->setTime($startHour, 0, 0);
+        $dayEndDateTime   = $dayStartDateTime->modify('+12 hours');
+
+        return $this->journeyModel
+            ->where('user_id', $userId)
+            ->where('start_datetime >=', $dayStartDateTime->format('Y-m-d H:i:s'))
+            ->where('start_datetime <',  $dayEndDateTime->format('Y-m-d H:i:s'))
+            ->first();
+    }  
 }
